fix(api): don't proxy EventEmitter
BREAKING CHANGE: This significant API change proxies `subscribe` and
`publish` instead of `on`, `once`, and `emit`.

There is no longer an equivalent for `once`.

This was done to keep compatibility with the "interface" of an
`EventEmitter`, among other things.

Also it will require Node.js v7 or newer due to use of features
unsupported in earlier versions.

Furthermore, it fixes connection issues arising from misuse of the
`pify` module.